+++ src/manager/recording/meson.build
@@ -1,0 +1,8 @@
+src_files += files(
+    'recording_reader.cpp',
+    'recording_reader.hpp',
+    'recording_writer.cpp',
+    'recording_writer.hpp',
+    'recording.cpp',
+    'recording.hpp',
+)
